+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - When editing nodes, you can now duplicate, duplicate with descendants, delete, delete with descendants with multiple nodes.
 - When editing nodes, you can now add UI elements to node layouts.
 - Added the ability to set the default placement of new nodes, when adding them to a navigation.
+- Added the ability to set the color for custom nodes and registered elements, which show as the color indicator on the type of node in the navigation builder.
 
 ### Changed
 - Redesigned UI for editing navigation nodes.
@@ -20,7 +21,7 @@
 - Custom URL nodes are now a Node Type - `verbb\navigation\nodetypes\CustomType`.
 
 ### Removed
-- Remove `NodeType::hasClasses()` and  `NodeType::hasAttributes()` which are now controlled by native fields.
+- Removed `NodeType::hasClasses()` and `NodeType::hasAttributes()` which are now controlled by native fields.
 
 ## 2.0.0-beta.2 - 2022-03-24